+++ code/Model/Grid3.h
@@ -159,7 +159,7 @@ public:
 		m_values = values;
 		m_indices.clear();
 
-		for (size_t i = 0; i < m_values.size(); ++i)
+		for (uint32_t i = 0; i < uint32_t(m_values.size()); ++i)
 		{
 			Vector4 p = PositionAccessor::get(m_values[i]) / m_cellSize;
 
@@ -185,7 +185,7 @@ public:
 
 	uint32_t size() const
 	{
-		return m_values.size();
+		return uint32_t(m_values.size());
 	}
 
 	const AlignedVector< ValueType >& values() const
